+++ agrest-annotations/src/main/java/io/agrest/annotation/LinkType.java
@@ -6,6 +6,7 @@ package io.agrest.annotation;
  * @since 2.10
  * @deprecated since 4.7, as Agrest now integrates with OpenAPI 3 / Swagger
  */
+@Deprecated
 public enum LinkType {
     COLLECTION, ITEM, METADATA, UNDEFINED
 }