+++ agrest-engine/src/main/java/io/agrest/AgResponse.java
@@ -2,24 +2,21 @@ package io.agrest;
 
 /**
  * A base response object in Agrest.
- * 
+ *
  * @since 1.19
  */
 public abstract class AgResponse {
-	
-	protected int status;
 
-	/**
-	 * @since 4.7
-	 */
-	public int getStatus() {
-		return status;
-	}
+    protected final int status;
 
-	/**
-	 * @since 4.7
-	 */
-	public void setStatus(int status) {
-		this.status = status;
-	}
+    public AgResponse(int status) {
+        this.status = status;
+    }
+
+    /**
+     * @since 4.7
+     */
+    public int getStatus() {
+        return status;
+    }
 }