+++ agrest-engine/src/test/java/io/agrest/DataResponseTest.java
@@ -1,22 +1,27 @@
 package io.agrest;
 
+import io.agrest.encoder.Encoder;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static java.util.Arrays.asList;
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.mock;
 
 public class DataResponseTest {
 
     @Test
-    public void testForType() {
-        DataResponse<Tr> response = DataResponse.forType(Tr.class);
-        assertNotNull(response);
+    public void testOf() {
 
         List<Tr> objects = asList(new Tr(), new Tr());
-        response.setObjects(objects);
-        assertEquals(objects, response.getObjects());
+        Encoder encoder = mock(Encoder.class);
+
+        DataResponse<Tr> response = DataResponse.of(201, objects, encoder);
+        assertNotNull(response);
+        assertEquals(201, response.getStatus());
+        assertSame(objects, response.getObjects());
+        assertSame(encoder, response.getEncoder());
     }
 
     public static class Tr {